powerpc/tm: Restore vs0 corruption
Currently tm_reclaim() returns with a corrupted vs0, due to the fact
that vs0 was used to restore VSCR.

Later, we recheckpoint from the live (not from the ckfp/ckvr) area,
depending on the MSR.

That said, we want to keep VEC and FP/VSX lives sane, because we might
want to possible recheckpoint from it.

Another fix is always forcing tm_recheckpoint() to recheckpoint everything
from the ckfp/ckvr area.