powerpc/tm: flush TM only if CPU has TM feature
Currently flush_tmregs_to_thread() does not check if TM feature is
available on CPU before trying to access TM SPRs in order to copy
live state to thread structures. flush_tmregs_to_thread() is indeed
guarded by CONFIG_PPC_TRANSACTIONAL_MEM but it might be the case
that kernel was compiled with CONFIG_PPC_TRANSACTIONAL_MEM enabled
and ran on a CPU without TM feature available, thus rendering the
execution of TM instructions that are treated by the CPU as
illegal instructions.

That commit fix this by checking in flush_tmregs_to_thread() if CPU has
the TM feature before accessing any TM-specific resource and returns
immediately if TM is no available on the CPU.

Signed-off-by: Gustavo Romero <gromero@linux.vnet.ibm.com>