powerpc/kernel: Always recheckpoint from ck state
Currently tm_recheckpoint() only restores part of the registers that are
explicitly set in the second argument. So, if a register set is not
specified (as Float Point registers), the current hot register values
will be recheckpoint in the CPU checkpoint area, not those from the
ckvec/ckfp values, which were reclaimed.

This is can cause some restore issues, as two examples:

 1) A task has FP enabled and traps into altivec_unavailable_tm(). This
will reclaim FP registers and put into ckfp state, but it will not
recheckpoint later because altivec_unavailable_tm() calls
tm_recheckpoint(&current->thread, MSR_VEC), which does not restore FP
from the ckfp state (just VEC), for FP, it restores from hot registers,
which might not contain the values pre tbegin.

A possible solution for this single case would be calling something as
tm_recheckpoint(&current->thread,  regs->msr), as regs->msr already has
MSR_FP set.

 2) On VSX unavailable exception on a task that contains FP enabled,
tm_reclaim will be called with the FP bit unset due to the regs->msr &
~orig_msr expression. This is not right, since FP will be restore from
the hot FP, which might contain an invalid value set by the tm_reclaim,
because it restored VEC, and used some fp registers during treclaim,
overwriting  vs0/fp0 with the content of VSCR.

The very same thing might happen if VEC is enabled but not FP. Then VEC
will be restored from VEC and it will be corrupted during FPSCR vs32
reclaim.

Both of these issues could be seen on the testcase that will follow this
patch.

To solve this issue, and simplify the whole reclaim/recheckpoint, I am
proposing to always recheckpoint from the checkpoint (ckvec/ckfp) area,
never from the live registers, so, if anything needs to be done before
recheckpoint, as, as filling out the lazy FP/VEC/VSX save/restore, it
needs to be done on the ckfp/ckvec area before the rechckpoint.

If that seems to be the right approach, tm_recheckpoint() could be
improved later to avoid the second argument at all.

Signed-off-by: Breno Leitao <leitao@debian.org>
Signed-off-by: Gustavo Romero <gusbromero@gmail.com>