selftests/powerpc: Check FP/VEC on exception in TM
Add a self test to check if FP/VEC/VSX registers are sane (restored
correctly) after a FP/VEC/VSX unavailable exception is caught in the
middle of a transaction.

This test checks all possibilities in a thread regarding the
MSR.[FP|VEC] state and for each case raises a FP/VEC/VSX unavailable
exception in transactional state, verifying if vs0 and vs32 registers,
which are representatives of FP/VEC/VSX reg sets, are not corrupted.

Signed-off-by: Gustavo Romero <gromero@linux.vnet.ibm.com>
Signed-off-by: Breno Leitao <leitao@debian.org>
Signed-off-by: Cyril Bur <cyrilbur@gmail.com>