selftests/powerpc: test for regs restore sanity
Add a self test to check if FP/VEC/VSX registers are sane (restored
correctly) after a FP/VEC/VSX unavailable exception is caught in the
middle of a transaction.

This patchset check all possibility of the thread, if it if has
FP/VEC/VSX enabled, and for each case, raise a VEC/VSX/FP unavailable_tm
exception.

Signed-off-by: Gustavo Romero <gromero@linux.vnet.ibm.com>
Signed-off-by: Breno Leitao <leitao@debian.org>