powerpc/tm: Remove msr_tm_active()
Currently msr_tm_active() is a wrapper around MSR_TM_ACTIVE() if
CONFIG_PPC_TRANSACTIONAL_MEM is set or a function that returns false
if CONFIG_PPC_TRANSACTIONAL_MEM is not set.

This function is not necessary since MSR_TM_ACTIVE() just check for some
bits and does not require any TM facility.

Signed-off-by: Breno Leitao <leitao@debian.org>