powerpc: Wire up fsconfig
Wire up a new syscall 'fsconfig' that was created by David and it is still not
on linux-next, but I need to add it in order to use the test case for the
previous other syscalls. I expect this syscall to be merged upstream also.

I tested this syscall on ppc64le (pseries) with the testcases from David
Howells and some other tests cases[2] that I built.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/dhowells/linux-fs.git/commit/?h=mount-context&id=2bae6fd63259c5244bd9de1b46cca706c4808438
[2] https://github.com/leitao/mcontext/

Signed-off-by: Breno Leitao <leitao@debian.org>