powerpc/tm: Always abort when entering kernel
Currently the syscall handler is not aborting when entering kernel in suspended
mode, only in transactional mode. Although the code documents that it checks if
the transaction is active (Transactional or Suspended), the check was only
against transactional.

This patch will abort a syscall if it is being called from inside a
transactional or suspended transaction.

This is causing a segfault on tm_sigreturn test case, where the signal handler
returns inside a suspended transaction.

Signed-off-by: Breno Leitao <leitao@debian.org>