powerpc/tm: Do not reclaim at syscall level
Current syscall code is tabort and returning to userspace if TM is active. We
should never reach this point if HTM was enabled in the signal handler and
returned, i.e, the signal handler starts a transaction and never conclude it.

There is one test[1] for this scenario, that returns in a resumed transaction.
This test is broken now, but I am wondering if it makes sense.

[1] tools/testing/selftests/powerpc/tm/tm-sigreturn.c

Signed-off-by: Breno Leitao <leitao@debian.org>