selftests/power: Redefine the test
The Documentation/powerpc/transactional_memory.txt says:

 "Syscalls made from within a suspended transaction are performed as normal
  and the transaction is not explicitly doomed by the kernel.  However,
  what the kernel does to perform the syscall may result in the transaction
  being doomed by the hardware."

The tm-syscall test is checking if the transaction is doomed or not, which
is an invalid proposition. With this new patchset, the syscall will
continue to execute, but the transaction will be doomed, since we will
reclaim and recheckpoint, so, I understand that we are not hurting the
documented behaviour above.

Signed-off-by: Breno Leitao <leitao@debian.org>