powerpc/tm: Wire up the syscall path
This is a patchset that invokes TM_KERNEL_ENTRY on the syscall entrance.

Since this macro will call a C function, we will need to restore the volatile
registers if the tm_reclaim end up being called.

In order to do so, TM_KERNEL_ENTRY returns 0 or 1, meaning if the tm_reclaim
code was executed (thus we need to restore the volatile registers) or if we can
just restore r0 and r3 (being used during the TM_KERNEL_ENTRY execution).

Signed-off-by: Breno Leitao <leitao@debian.org>