selftests/powerpc: Do not fail on TM_CAUSE_RESCHED
There are cases where we are not expecting to have the transaction aborted,
but, the test process might have been rescheduled, either in the OS level or by
KVM if it is running on a KVM guest machine. The process reschedule will cause
a treclaim/recheckpoint which will cause the transaction to doom, failing as
soon as the process is rescheduled.

If that is the case, TEXASR[FC] is indicated with either TM_CAUSE_RESCHEDULE or
TM_CAUSE_KVM_RESCHEDULE for KVM interruptions.

In this scenario, disconsider it and avoid the test to return failure

Signed-off-by: Breno Leitao <leitao@debian.org>