powerpc/kernel: Fix potential spectre v1 in syscall
The rtas syscall reads a value from a user-provided structure and uses it to
index an array, being a possible area for a possible spectre v1 attack. This is
the code that has this issue.

	args.rets = &args.args[nargs];

The nargs is a userprovided value, and the below code would set nargs to XX.

	struct rtas_args ra;
	ra.nargs = htobe32(XX);
	syscall(__NR_rtas, &ra);

Signed-off-by: Breno Leitao <leitao@debian.org>