powerpc/tm: Create a macro that reclaims
This is the macro that will be invoked on all entrance to the kernel, so, in
kernel space the transaction will be completely reclaimed and not suspended
anymore.

This patchset checks if we are coming from PR, if not, skip. This is useful
when there is a irq_replay() being called after recheckpoint, when the IRQ is
re-enable. In this case, we do not want to re-reclaim and re-recheckpoint,
thus, if not coming from PR, skip it completely.

This macro does not care about TM SPR also, it will only be saved and restore
in the context switch code now on.

This macro will return 0 or 1 in r3 register, to specify if a reclaim was
executed or not.

This patchset is based on inital work done by Cyril:
https://patchwork.ozlabs.org/cover/875341/

Signed-off-by: Breno Leitao <leitao@debian.org>