powerpc/tm: Restructure the __switch_to code
This is the code that will switch between two tasks TM-wise. It is clearly
split in two parts, the task that is leaving the CPU, known as prev and the
task that is being scheduled, known as new.

It starts checking if the previous task had TM enable, if so, it increases
the load_tm (this is the only place we increment load_tm). It also saves
the SPRs here.

If the previous task was scheduled out with a transaction active, we need
to fix the failure cause, since it might contain the failure cause that
caused the exception, as TM_CAUSE_MISC. In this case, since there was a
context switch, overwrite the failure cause.

If the previous task has overflowed load_tm, disable TM also. Even if
TIF_RESTORE_TM is on, it will be not reload the the SPRs anymore,
restarting the lazy save/restore mechanism.

When loading the new task, it basically restore the SPRs, and
TIF_RESTORE_TM (set by tm_reclaim_current if the transaction was active)
would invoke the recheckpoint process later in restore_tm_state().

On top of that, both tm_reclaim_task() and tm_recheckpoint_new_task()
functions are not used anymore, removing them.

Signed-off-by: Breno Leitao <leitao@debian.org>