powerpc/tm: Use TM_KERNEL_ENTRY on unvailable_tm
If there is a FP/VEC/Altivec touch inside a transaction, it called the
xx_unavailable_tm, which was reclaiming and recheckpointing. That is not required
anymore, since the code was reclaimed in the trap itself, and it will be
recheckpointed by restore_tm_state.

With this new patchset, we only reclaim at exception entrance (except if
sigreturn returns suspended), and always recheckpoint on restore_tm_state.

Adding a WARN_ON() warning if we hit the _unavailable_tm() in suspended mode,
i.e, the reclaim was not executed somehow.

Signed-off-by: Breno Leitao <leitao@debian.org>