powerpc/tm: recheckpoint on our way to userspace
This is the only place we are going to recheckpoint now. I.e, the task need
to have TIF_RESTORE_TM which will get into restore_tm_state() which will
recheckpoint if MSR shows that the transaction was active.

Every time a task is required to recheckpoint, or or just restore the TM
SPRS, TIF_RESTORE_TM should be set.

Signed-off-by: Breno Leitao <leitao@debian.org>