powerpc/tm: Function that updates the failure code
Now the transaction reclaims happens very earlier in the trap handler, and
it is impossible to know, at that early time, what should be set as the
failure case for some specific cases, as, if the transaction will be
rescheduled, thus, the abort case should be updated from TM_CAUSE_MISC to
TM_CAUSE_RESCHED, for example).

This patch provides a function that will update TEXAS special purpose
register int the task and set the failure code which will be discovered
afterward.

Signed-off-by: Breno Leitao <leitao@debian.org>