powerpc/tm: Do not recheckpoint at sigreturn
Do not recheckpoint at signal code return. Just make sure TIF_RESTORE_TM is
set, which will restore in the exit to userspace by restore_tm_state.

All the FP and VEC lazy restore was already done by tm_reclaim_current(),
where it check if FP/VEC was set, and set the ckfp and ckvr registers to
the expected value. The current FP/VEC restoration is not necessary, since
the transaction will be aborted and the checkpointed values will be
restore.

Signed-off-by: Breno Leitao <leitao@debian.org>