powerpc/tm: Do not restore default DSCR
Since treclaim just happens before getting userspace, we do not need to restore
DSCR to the default value. We can keep the checkpointed one, mainly because the
transaction is doommed and it will fail.

Without this change, userspace is returning with the default DSCR value,
independentely of the previous DSCR value set.

Signed-off-by: Breno Leitao <leitao@debian.org>