powerpc/tm: Recheckpoint later
We were currently recheckpointing with the rest of the TIF workers, but,
that was too early, and can cause some IRQ to be replayed.

This patch moves the tm_recheckpoint to the last possible place, and
force the IRQ to be disabled, and it will continue to be until RFID, so,
no IRQ will be replayed at all.

Also TIF_RESTORE_TM is now a special case that should not be handled
similarly to the _TIF_USER_WORK_MASK.

Since _TIF_RESTORE_TM is not part of _TIF_USER_WORK_MASK anymore, we
need to force system_call_exit to continue to leaves through
fast_exception_return, so, we add the flags together with
_TIF_USER_WORK_MASK at system_call_exist path.

If this flag is set at system_call_exist, it means that recheckpoint
will be called, and doing it through fast_exception_return is the only
way to do so.

Signed-off-by: Breno Leitao <leitao@debian.org>